Move src to src/nonemast
This is slightly more standard layout:
https://github.com/pypa/sampleproject
https://packaging.python.org/en/latest/tutorials/packaging-projects/?highlight=src#a-simple-project
https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/
https://setuptools.pypa.io/en/latest/userguide/package_discovery.html#src-layout

It should allow us to run nonemast without installation by pointing `PYTHONPATH` to `src/`.
