+++ src/nonemast/update-details.ui
